addressed issues in code review
